+++ src/super/i-block/modules/dom/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.?? ()
+
+#### :boom: Breaking Change
+
+* Now `localInView` returns a promise
+
 ## v3.0.0-rc.60 (2020-09-01)
 
 #### :rocket: New Feature